breakup monolithic contributions.views module
