+++ apps/contributions/tests/test_views.py
@@ -1,3 +1,5 @@
+# TODO @BW: submodularize this file to align with views module structure
+# DEV-5536
 import datetime
 import json
 from unittest import mock
@@ -21,7 +23,6 @@ from waffle import get_waffle_flag_model
 
 from apps.common.constants import CONTRIBUTIONS_API_ENDPOINT_ACCESS_FLAG_NAME
 from apps.common.tests.test_resources import AbstractTestCase
-from apps.contributions import views as contributions_views
 from apps.contributions.models import (
     Contribution,
     ContributionInterval,
@@ -40,6 +41,7 @@ from apps.contributions.tests.factories import (
     ContributorFactory,
     PaymentFactory,
 )
+from apps.contributions.views import portal as portal_views
 from apps.organizations.tests.factories import (
     OrganizationFactory,
     RevenueProgramFactory,
@@ -79,7 +81,7 @@ class StripeOAuthTest(AbstractTestCase):
             body["scope"] = scope
         return self.client.post(complete_url, body)
 
-    @mock.patch("apps.contributions.views.task_verify_apple_domain")
+    @mock.patch("apps.contributions.views.orgs.task_verify_apple_domain")
     @mock.patch("stripe.OAuth.token")
     def test_response_when_missing_params(self, stripe_oauth_token, task_verify_apple_domain):
         # Missing code
@@ -107,7 +109,7 @@ class StripeOAuthTest(AbstractTestCase):
         stripe_oauth_token.assert_not_called()
         assert not task_verify_apple_domain.delay.called
 
-    @mock.patch("apps.contributions.views.task_verify_apple_domain")
+    @mock.patch("apps.contributions.views.orgs.task_verify_apple_domain")
     @mock.patch("stripe.OAuth.token")
     def test_response_when_scope_param_mismatch(self, stripe_oauth_token, task_verify_apple_domain):
         """We verify that the "scope" parameter provided by the frontend matches the scope we expect."""
@@ -117,7 +119,7 @@ class StripeOAuthTest(AbstractTestCase):
         stripe_oauth_token.assert_not_called()
         assert not task_verify_apple_domain.delay.called
 
-    @mock.patch("apps.contributions.views.task_verify_apple_domain")
+    @mock.patch("apps.contributions.views.orgs.task_verify_apple_domain")
     @mock.patch("stripe.OAuth.token")
     def test_response_when_invalid_code(self, stripe_oauth_token, task_verify_apple_domain):
         stripe_oauth_token.side_effect = StripeInvalidGrantError(code="error_code", description="error_description")
@@ -127,7 +129,7 @@ class StripeOAuthTest(AbstractTestCase):
         stripe_oauth_token.assert_called_with(code="1234", grant_type="authorization_code")
         assert not task_verify_apple_domain.delay.called
 
-    @mock.patch("apps.contributions.views.task_verify_apple_domain")
+    @mock.patch("apps.contributions.views.orgs.task_verify_apple_domain")
     @mock.patch("stripe.OAuth.token")
     def test_response_success(self, stripe_oauth_token, task_verify_apple_domain):
         expected_stripe_account_id = "my_test_account_id"
@@ -147,7 +149,7 @@ class StripeOAuthTest(AbstractTestCase):
         assert Version.objects.get_for_object(self.org1_rp1.payment_provider).count() == 1
         task_verify_apple_domain.delay.assert_called_with(revenue_program_slug=self.org1_rp1.slug)
 
-    @mock.patch("apps.contributions.views.task_verify_apple_domain")
+    @mock.patch("apps.contributions.views.orgs.task_verify_apple_domain")
     @mock.patch("stripe.OAuth.token")
     def test_create_payment_provider_if_not_exists(self, stripe_oauth_token, task_verify_apple_domain):
         expected_stripe_account_id = "new_stripe_account_id"
@@ -568,16 +570,16 @@ class TestProcessStripeWebhook:
 
     def test_happy_path(self, api_client, mocker):
         mocker.patch("stripe.Webhook.construct_event", return_value=(event := mocker.Mock()))
-        mock_process_task = mocker.patch("apps.contributions.views.process_stripe_webhook_task.delay")
+        mock_process_task = mocker.patch("apps.contributions.views.webhooks.process_stripe_webhook_task.delay")
         header = {"HTTP_STRIPE_SIGNATURE": "testing", "content_type": "application/json"}
         response = api_client.post(reverse("stripe-webhooks-contributions"), data={}, **header)
         assert response.status_code == status.HTTP_200_OK
         mock_process_task.assert_called_once_with(raw_event_data=event)
 
     def test_when_value_error_on_construct_event(self, api_client, mocker):
-        logger_spy = mocker.patch("apps.contributions.views.logger.warning")
+        logger_spy = mocker.patch("apps.contributions.views.webhooks.logger.warning")
         mocker.patch("stripe.Webhook.construct_event", side_effect=ValueError("ruh roh"))
-        mock_process_task = mocker.patch("apps.contributions.views.process_stripe_webhook_task.delay")
+        mock_process_task = mocker.patch("apps.contributions.views.webhooks.process_stripe_webhook_task.delay")
         header = {"HTTP_STRIPE_SIGNATURE": "testing", "content_type": "application/json"}
         response = api_client.post(reverse("stripe-webhooks-contributions"), data={"foo": "bar"}, **header)
         assert response.status_code == status.HTTP_400_BAD_REQUEST
@@ -585,11 +587,11 @@ class TestProcessStripeWebhook:
         logger_spy.assert_called_once_with("Invalid payload from Stripe webhook request")
 
     def test_when_signature_verification_error(self, api_client, mocker):
-        logger_spy = mocker.patch("apps.contributions.views.logger.exception")
+        logger_spy = mocker.patch("apps.contributions.views.webhooks.logger.exception")
         mocker.patch(
             "stripe.Webhook.construct_event", side_effect=stripe.error.SignatureVerificationError("ruh roh", "sig")
         )
-        mock_process_task = mocker.patch("apps.contributions.views.process_stripe_webhook_task.delay")
+        mock_process_task = mocker.patch("apps.contributions.views.webhooks.process_stripe_webhook_task.delay")
         header = {"HTTP_STRIPE_SIGNATURE": "testing", "content_type": "application/json"}
         response = api_client.post(reverse("stripe-webhooks-contributions"), data={"foo": "bar"}, **header)
         assert response.status_code == status.HTTP_400_BAD_REQUEST
@@ -1552,7 +1554,7 @@ class TestPortalContributorsViewSet:
             ContributionQuerySet, "exclude_recurring_missing_provider_subscription_id"
         )
         exclude_dummy_payment_method_id = mocker.spy(ContributionQuerySet, "exclude_dummy_payment_method_id")
-        contributions_views.PortalContributorsViewSet().get_contributor_queryset(contributor=ContributorFactory())
+        portal_views.PortalContributorsViewSet().get_contributor_queryset(contributor=ContributorFactory())
         exclude_hidden_spy.assert_called_once()
         exclude_paymentless_spy.assert_called_once()
         exclude_missing_stripe_sub_id.assert_called_once()
