+++ apps/contributions/tests/test_webhooks_integration.py
@@ -451,7 +451,7 @@ def test_invoice_upcoming(interval, expect_reminder_email, contribution_found, c
 
 def test_process_stripe_webhook_when_value_error_raised(mocker, client):
     mocker.patch("stripe.Webhook.construct_event", side_effect=ValueError("ruh roh"))
-    logger_spy = mocker.patch("apps.contributions.views.logger.warning")
+    logger_spy = mocker.patch("apps.contributions.views.webhooks.logger.warning")
     header = {"HTTP_STRIPE_SIGNATURE": "testing", "content_type": "application/json"}
     response = client.post(reverse("stripe-webhooks-contributions"), data={}, **header)
     assert response.status_code == status.HTTP_400_BAD_REQUEST
