+++ apps/contributions/views/orgs.py
@@ -1,0 +1,179 @@
+"""Contains views for API resources that are exposed to organizations."""
+
+import logging
+
+from django.conf import settings
+from django.db.models import QuerySet
+
+import stripe
+from django_filters.rest_framework import DjangoFilterBackend
+from rest_framework import status, viewsets
+from rest_framework.decorators import action, api_view, permission_classes
+from rest_framework.permissions import IsAdminUser, IsAuthenticated
+from rest_framework.response import Response
+from reversion.views import create_revision
+
+from apps.api.exceptions import ApiConfigurationError
+from apps.api.permissions import (
+    HasFlaggedAccessToContributionsApiResource,
+    HasRoleAssignment,
+    IsContributor,
+)
+from apps.contributions import serializers
+from apps.contributions.filters import ContributionFilter
+from apps.contributions.models import Contribution
+from apps.contributions.tasks import (
+    email_contribution_csv_export_to_user,
+    task_verify_apple_domain,
+)
+from apps.organizations.models import PaymentProvider, RevenueProgram
+from apps.public.permissions import IsActiveSuperUser
+from apps.users.models import User
+
+
+logger = logging.getLogger(f"{settings.DEFAULT_LOGGER}.{__name__}")
+
+
+@create_revision()
+@api_view(["POST"])
+@permission_classes([IsAuthenticated, HasRoleAssignment | IsActiveSuperUser])
+def stripe_oauth(request):
+    scope = request.data.get("scope")
+    code = request.data.get("code")
+    revenue_program_id = request.data.get("revenue_program_id")
+    if not revenue_program_id:
+        return Response(
+            {"missing_params": "revenue_program_id missing required params"}, status=status.HTTP_400_BAD_REQUEST
+        )
+
+    if not scope or not code:
+        return Response({"missing_params": "stripe_oauth missing required params"}, status=status.HTTP_400_BAD_REQUEST)
+
+    if scope != settings.STRIPE_OAUTH_SCOPE:
+        return Response(
+            {"scope_mismatch": "stripe_oauth received unexpected scope"}, status=status.HTTP_400_BAD_REQUEST
+        )
+
+    revenue_program = RevenueProgram.objects.get(id=revenue_program_id)
+
+    try:
+        logger.info("[stripe_oauth] Attempting to perform oauth with Stripe for revenue program %s", revenue_program_id)
+        oauth_response = stripe.OAuth.token(
+            grant_type="authorization_code",
+            code=code,
+        )
+        payment_provider = revenue_program.payment_provider
+        if not payment_provider:
+            logger.info(
+                "[stripe_oauth] Payment provider not yet created for revenue program %s, proceeding with creation...",
+                revenue_program_id,
+            )
+            payment_provider = PaymentProvider.objects.create(
+                stripe_account_id=oauth_response["stripe_user_id"],
+                stripe_oauth_refresh_token=oauth_response["refresh_token"],
+            )
+            logger.info(
+                "[stripe_oauth] Payment provider for revenue program %s created: %s",
+                revenue_program_id,
+                payment_provider.id,
+            )
+            revenue_program.payment_provider = payment_provider
+            revenue_program.save()
+        else:
+            logger.info(
+                "[stripe_oauth] Updating existing payment provider %s for revenue program %s",
+                payment_provider.id,
+                revenue_program_id,
+            )
+            payment_provider.stripe_account_id = oauth_response["stripe_user_id"]
+            payment_provider.stripe_oauth_refresh_token = oauth_response["refresh_token"]
+            payment_provider.save()
+        logger.info(
+            "[stripe_oauth] Starting Apple Pay domain verification background task for revenue program slug: %s",
+            revenue_program.slug,
+        )
+        task_verify_apple_domain.delay(revenue_program_slug=revenue_program.slug)
+        logger.info(
+            "[stripe_oauth] Started Apple Pay domain verification background task for revenue program slug: %s",
+            revenue_program.slug,
+        )
+    except stripe.oauth_error.InvalidGrantError:
+        logger.exception("[stripe_oauth] stripe.OAuth.token failed due to an invalid code")
+        return Response({"invalid_code": "stripe_oauth received an invalid code"}, status=status.HTTP_400_BAD_REQUEST)
+
+    return Response({"detail": "success"}, status=status.HTTP_200_OK)
+
+
+class ContributionsViewSet(viewsets.ReadOnlyModelViewSet):
+    """Contributions API resource.
+
+    NB: There are bespoke actions on this viewset that override the default permission classes set here.
+    """
+
+    permission_classes = [
+        IsAuthenticated,
+        (HasFlaggedAccessToContributionsApiResource & (HasRoleAssignment | IsActiveSuperUser)),
+    ]
+    model = Contribution
+    filterset_class = ContributionFilter
+    filter_backends = [DjangoFilterBackend]
+    serializer_class = serializers.ContributionSerializer
+
+    def filter_queryset_for_user(self, user: User) -> QuerySet[Contribution]:
+        """Return the right results to the right user."""
+        ra = getattr(user, "get_role_assignment", lambda: None)()
+        if user.is_anonymous:
+            return self.model.objects.none()
+        if user.is_superuser:
+            return self.model.objects.all()
+        if ra:
+            return self.model.objects.filtered_by_role_assignment(ra)
+        logger.warning("Encountered unexpected user %s", user.id)
+        raise ApiConfigurationError
+
+    def get_queryset(self):
+        """Return the right results to the right user."""
+        ra = getattr((user := self.request.user), "get_role_assignment", lambda: None)()
+        if user.is_anonymous:
+            return self.model.objects.none()
+        if user.is_superuser:
+            return self.model.objects.all()
+        if ra:
+            return self.model.objects.filtered_by_role_assignment(ra)
+        logger.warning("Encountered unexpected user %s", user.id)
+        raise ApiConfigurationError
+
+    @action(
+        methods=["post"],
+        url_path="email-contributions",
+        detail=False,
+        permission_classes=[~IsContributor, IsAuthenticated, IsAdminUser | HasRoleAssignment],
+    )
+    def email_contributions(self, request):
+        """Endpoint to send contributions as a csv file to the user request.
+
+        Any user who has role and authenticated will be able to call the endpoint.
+        Contributor will not be able to access this endpoint as it's being integrated with the Contribution Dashboard
+        as contributors will be able to access only Contributor Portal via magic link.
+        """
+        logger.info(
+            "enqueueing email_contribution_csv_export_to_user task for user %s",
+            request.user,
+        )
+        logger.debug(
+            "enqueueing email_contribution_csv_export_to_user task for request: %s",
+            request,
+        )
+        ra = request.user.get_role_assignment()
+        show_upgrade_prompt = (
+            not request.user.is_superuser
+            and ra
+            and (org := getattr(ra, "organization", None))
+            and org.plan.name == "FREE"
+        )
+        email_contribution_csv_export_to_user.delay(
+            list(self.get_queryset().values_list("id", flat=True)),
+            request.user.email,
+            show_upgrade_prompt,
+        )
+        return Response(data={"detail": "success"}, status=status.HTTP_200_OK)
